+++ src/resource/solidDataset.ts
@@ -26,7 +26,8 @@ import {
   getChainBlankNodes,
   toRdfJsQuads,
 } from "../rdfjs.internal";
-import { ldp } from "../constants";
+import { ldp, pim } from "../constants";
+import { getJsonLdParser } from "../formats/jsonLd";
 import { triplesToTurtle, getTurtleParser } from "../formats/turtle";
 import { isLocalNode, isNamedNode, resolveIriForLocalNode } from "../datatypes";
 import {
@@ -42,6 +43,7 @@ import {
   WithChangeLog,
   hasChangelog,
   LocalNode,
+  SolidClientError,
 } from "../interfaces";
 import { internal_toIriString } from "../interfaces.internal";
 import {
@@ -52,6 +54,7 @@ import {
   FetchError,
   responseToResourceInfo,
   getContentType,
+  getLinkedResourceUrlAll,
 } from "./resource";
 import {
   internal_isUnsuccessfulResponse,
@@ -1133,5 +1136,41 @@ function resolveLocalIrisInThing(
     url: isLocalNodeIri(thing.url)
       ? `${baseIri}#${getLocalNodeName(thing.url)}`
       : thing.url,
+  });
+}
+
+/**
+ * Fetch the contents of '.well-known/solid' for a given resource URL.
+ *
+ * The contents of the '.well-known/solid' endpoint define the capabilities of the server, and provide their associated endpoints/locations.
+ * This behaves similarly to the use of '.well-known' endpoints in e.g. (OIDC servers)[https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfig]
+ *
+ * @param url URL of a Resource.
+ * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/docs/Web/API/WindowOrWorkerGlobalScope/fetch#parameters).
+ * @returns Promise resolving to a [[SolidDataset]] containing the data at '.well-known/solid' for the given Resource, or rejecting if fetching it failed.
+ */
+export async function getWellKnownSolid(
+  url: UrlString | Url,
+  options: Partial<
+    typeof internal_defaultFetchOptions & ParseOptions
+  > = internal_defaultFetchOptions
+): Promise<SolidDataset & WithServerResourceInfo> {
+  const urlString = internal_toIriString(url);
+  const resourceMetadata = await getResourceInfo(urlString, options);
+  const linkedResources = getLinkedResourceUrlAll(resourceMetadata);
+  const rootResources = linkedResources[pim.storage];
+  const rootResource = rootResources?.length === 1 ? rootResources[0] : null;
+  if (rootResource === null) {
+    throw new SolidClientError(
+      `Unable to determine root resource for Resource at [${url}].`
+    );
+  }
+  const wellKnownSolidUrl = new URL(".well-known/solid", rootResource).href;
+
+  return getSolidDataset(wellKnownSolidUrl, {
+    ...options,
+    parsers: {
+      "application/ld+json": getJsonLdParser(),
+    },
   });
 }